simplify partition size reading.
